SNOW-1727163: Do not drop table in temp table cleaner if the session is already closed (#2409)
<!---
Please answer these questions before creating your pull request. Thanks!
--->

1. Which Jira issue is this PR addressing? Make sure that there is an
accompanying issue to your PR.

   <!---
   In this section, please add a Snowflake Jira issue number.
   
Note that if a corresponding GitHub issue exists, you should still
include
   the Snowflake Jira issue number. For example, for GitHub issue
https://github.com/snowflakedb/snowpark-python/issues/1400, you should
   add "SNOW-1335071" here.
    --->

   Fixes SNOW-1727163

2. Fill out the following pre-review checklist:

- [x] I am adding a new automated test(s) to verify correctness of my
new code
- [ ] If this test skips Local Testing mode, I'm requesting review from
@snowflakedb/local-testing
   - [ ] I am adding new logging messages
   - [ ] I am adding a new telemetry message
   - [ ] I am adding new credentials
   - [ ] I am adding a new dependency
- [ ] If this is a new feature/behavior, I'm adding the Local Testing
parity changes.

3. Please describe how your code solves the related issue.

It's possible that the session is already closed before garbage
collection kicks in, where we should avoid sending drop table sql and
eliminate the warning

+++ src/snowflake/snowpark/_internal/temp_table_auto_cleaner.py
@@ -45,7 +45,12 @@ class TempTableAutoCleaner:
         """
         self.ref_count_map[name] -= 1
         if self.ref_count_map[name] == 0:
-            if self.session.auto_clean_up_temp_table_enabled:
+            if (
+                self.session.auto_clean_up_temp_table_enabled
+                # if the session is already closed before garbage collection,
+                # we have no way to drop the table
+                and not self.session._conn.is_closed()
+            ):
                 self.drop_table(name)
         elif self.ref_count_map[name] < 0:
             logging.debug(
@@ -80,12 +85,13 @@ class TempTableAutoCleaner:
         """
         Stops the cleaner (no-op) and sends the telemetry.
         """
-        self.session._conn._telemetry_client.send_temp_table_cleanup_telemetry(
-            self.session.session_id,
-            temp_table_cleaner_enabled=self.session.auto_clean_up_temp_table_enabled,
-            num_temp_tables_cleaned=self.num_temp_tables_cleaned,
-            num_temp_tables_created=self.num_temp_tables_created,
-        )
+        if not self.session._conn.is_closed():
+            self.session._conn._telemetry_client.send_temp_table_cleanup_telemetry(
+                self.session.session_id,
+                temp_table_cleaner_enabled=self.session.auto_clean_up_temp_table_enabled,
+                num_temp_tables_cleaned=self.num_temp_tables_cleaned,
+                num_temp_tables_created=self.num_temp_tables_created,
+            )
 
     @property
     def num_temp_tables_created(self) -> int:
